+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- 新增： fsext.py 加入 get_files_glob 函数，以替代原来的 get_files 函数。
+
 # [0.3.9] 23-02-16
 
 - 修改： autoupgrade 指定 upgrade version，并加入 force_reinstall 参数。